make it build

+++ src/WasteTrader/Database/IWaste.cs
@@ -7,8 +7,8 @@ namespace WasteTrader.Database
         string Name { get; set; }
         string Description { get; set; }
         long Quantity { get; set; }
-        sbyte UnitMetricPrefixPower { get; set; }
-        byte Unit { get; set; }
+        int UnitMetricPrefixPower { get; set; }
+        int Unit { get; set; }
         long Price { get; set; }
         DateTime EntryTime { get; }
         int Category { get; set; }
