+++ src/WasteTrader/Measurements/Area.cs
@@ -5,24 +5,24 @@ namespace WasteTrader.Measurements
 {
     public class Area : Measurement<Area>
     {
-        public Area(long Quantity, sbyte UnitMetricPrefixPower)
+        public Area(long Quantity, int UnitMetricPrefixPower)
         {
             this.Quantity = Quantity;
             this.UnitMetricPrefixPower = UnitMetricPrefixPower;
         }
 
-        private static ImmutableDictionary<sbyte, Unit> Units = MetricPrefixes.Symbol.ToImmutableDictionary(ConvertKey, ConvertValue);
+        private static ImmutableDictionary<int, Unit> Units = MetricPrefixes.Symbol.ToImmutableDictionary(ConvertKey, ConvertValue);
 
-        private static sbyte ConvertKey(KeyValuePair<sbyte, string> pair)
+        private static int ConvertKey(KeyValuePair<int, string> pair)
         {
-            return (sbyte)(pair.Key * 2);
+            return (int)(pair.Key * 2);
         }
 
-        private static Unit ConvertValue(KeyValuePair<sbyte, string> pair)
+        private static Unit ConvertValue(KeyValuePair<int, string> pair)
         {
             return new Unit(pair.Value + "m²", 0);
         }
 
-        public override IImmutableDictionary<sbyte, Unit> Symbols => Units;
+        public override IImmutableDictionary<int, Unit> Symbols => Units;
     }
 }
