+++ src/WasteTrader/Measurements/Length.cs
@@ -5,25 +5,25 @@ namespace WasteTrader.Measurements
 {
     public class Length : Measurement<Length>
     {
-        public Length(long Quantity, sbyte UnitMetricPrefixPower)
+        public Length(long Quantity, int UnitMetricPrefixPower)
         {
             this.Quantity = Quantity;
             this.UnitMetricPrefixPower = UnitMetricPrefixPower;
         }
 
-        private static ImmutableDictionary<sbyte, Unit> Units = MetricPrefixes.Symbol.ToImmutableDictionary(ConvertKey, ConvertValue);
+        private static ImmutableDictionary<int, Unit> Units = MetricPrefixes.Symbol.ToImmutableDictionary(ConvertKey, ConvertValue);
 
-        private static sbyte ConvertKey(KeyValuePair<sbyte, string> pair)
+        private static int ConvertKey(KeyValuePair<int, string> pair)
         {
             return pair.Key;
         }
 
-        private static Unit ConvertValue(KeyValuePair<sbyte, string> pair)
+        private static Unit ConvertValue(KeyValuePair<int, string> pair)
         {
             return new Unit(pair.Value + "m", 0);
         }
 
 
-        public override IImmutableDictionary<sbyte, Unit> Symbols => Units;
+        public override IImmutableDictionary<int, Unit> Symbols => Units;
     }
 }
