+++ src/WasteTrader/Measurements/Unit.cs
@@ -2,12 +2,12 @@
 {
     public class Unit
     {
-        public Unit(string Text, sbyte Offset)
+        public Unit(string Text, int Offset)
         {
             this.Text = Text;
             this.Offset = Offset;
         }
         public string Text { get; }
-        public sbyte Offset { get; }
+        public int Offset { get; }
     }
 }
